# AurigaMail/src/main/webapp/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Espone servizi REST per le mail",
    "version" : "v1",
    "title" : "Servizi di AurigaMail",
    "termsOfService" : "",
    "contact" : {
      "name" : "API Support",
      "url" : "http://www.eng.it/",
      "email" : "Ugo.PiconeChiodo@eng.it"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host" : "",
  "basePath" : "/AurigaMail/rest",
...
  "paths" : {
...
    "/email/{emailId}/attachments" : {
      "get" : {
        "tags" : [ "Interazione" ],
        "summary" : "Recupera gli allegati dell'email (RecuperaAllegatiEmail)",
        "description" : "Chiama ServiceRestStore.invoke() del core-module (nuova operazione 'getAttachmentsWithContents').",
        "operationId" : "getAttachmentsWithContents",
        "produces" : [ "application/xml" ],
        "parameters" : [ {
          "name" : "emailId",
          "in" : "path",
          "description" : "ID email",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "flagGetContent",
          "in" : "query",
          "description" : "se false non recupera il contenuto",
          "required" : false,
          "type" : "boolean",
          "default" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/GetEmailAttachmentsResponse"
            }
          },
          "400" : {
            "description" : "Operazione interrotta a causa di input non valido."
          },
          "404" : {
            "description" : "Operazione interrotta a causa di dati non trovati e/o non validi."
          },
          "500" : {
            "description" : "Operazione interrotta a causa di un errore imprevisto."
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "EmailAttachment" : {
      "type" : "object",
      "properties" : {
        "messageid" : {
          "type" : "string",
          "xml" : {
            "name" : "idMessaggio"
          }
        },
        "filename" : {
          "type" : "string",
          "xml" : {
            "name" : "nomeFile"
          }
        },
        "mimetype" : {
          "type" : "string",
          "xml" : {
            "name" : "tipoMIME"
          }
        },
        "contentID" : {
          "type" : "string",
          "xml" : {
            "name" : "idContenuto"
          }
        },
        "disposition" : {
          "type" : "string",
          "xml" : {
            "name" : "disposizione"
          }
        },
        "size" : {
          "type" : "integer",
          "format" : "int64",
          "xml" : {
            "name" : "dimensione"
          }
        },
        "contents" : {
          "type" : "array",
          "xml" : {
            "name" : "contenuto"
          },
          "items" : {
            "type" : "string",
            "format" : "byte",
            "xml" : {
              "name" : "contenuto"
            }
          }
        },
        "encodedHash" : {
          "type" : "string",
          "xml" : {
            "name" : "hashCodificato"
          }
        },
        "encoding" : {
          "type" : "string",
          "xml" : {
            "name" : "codifica"
          }
        },
        "algoritmo" : {
          "type" : "string"
        }
      },
      "xml" : {
        "name" : "allegatoEmail"
      }
    },
...
    "GetEmailAttachmentsResponse" : {
      "type" : "object",
      "properties" : {
        "mailAttachments" : {
          "type" : "array",
          "xml" : {
            "name" : "allegatoEmail"
          },
          "items" : {
            "xml" : {
              "name" : "allegatoEmail"
            },
            "$ref" : "#/definitions/EmailAttachment"
          }
        },
        "files" : {
          "type" : "array",
          "xml" : {
            "name" : "file"
          },
          "items" : {
            "type" : "file",
            "xml" : {
              "name" : "file"
            }
          }
        }
      },
      "xml" : {
        "name" : "rispostaRecuperaAllegatiEmail"
      }
    },
...
  }
}